+++ test/extensions/filters/http/lua/lua_integration_test.cc
@@ -109,6 +109,7 @@ config:
       request_handle:headers():add("request_body_size", body_length)
       request_handle:headers():add("request_metadata_foo", metadata["foo"])
       request_handle:headers():add("request_metadata_baz", metadata["baz"])
+      request_handle:headers():add("request_protocol", request_handle:requestInfo():protocol())
     end
 
     function envoy_on_response(response_handle)
@@ -117,6 +118,7 @@ config:
       response_handle:headers():add("response_metadata_foo", metadata["foo"])
       response_handle:headers():add("response_metadata_baz", metadata["baz"])
       response_handle:headers():add("response_body_size", body_length)
+      response_handle:headers():add("request_protocol", response_handle:requestInfo():protocol())
       response_handle:headers():remove("foo")
     end
 )EOF";
@@ -153,6 +155,10 @@ config:
                           ->value()
                           .c_str());
 
+  EXPECT_STREQ(
+      "HTTP/1.1",
+      upstream_request_->headers().get(Http::LowerCaseString("request_protocol"))->value().c_str());
+
   Http::TestHeaderMapImpl response_headers{{":status", "200"}, {"foo", "bar"}};
   upstream_request_->encodeHeaders(response_headers, false);
   Buffer::OwnedImpl response_data1("good");
@@ -170,6 +176,8 @@ config:
   EXPECT_STREQ(
       "bat",
       response->headers().get(Http::LowerCaseString("response_metadata_baz"))->value().c_str());
+  EXPECT_STREQ("HTTP/1.1",
+               response->headers().get(Http::LowerCaseString("request_protocol"))->value().c_str());
   EXPECT_EQ(nullptr, response->headers().get(Http::LowerCaseString("foo")));
 
   cleanup();